Consolidate export pipeline around PublishTarget interface
Lifts the export pipeline out of foam-cli into foam-core and refactors
it around a PublishTarget interface.

Two targets land here:
- Starlight (refactored from the existing CLI exporter onto the new interface)
- HTML-page (bridge over renderReport for the VS Code "Export to HTML page" command)
